support for HA tests
