+++ phoenix-core/src/it/java/org/apache/phoenix/end2end/PhoenixRegionServerEndpointHAImpl.java
@@ -1,0 +1,11 @@
+package org.apache.phoenix.end2end;
+
+import org.apache.phoenix.cache.ServerMetadataCache;
+import org.apache.phoenix.coprocessor.PhoenixRegionServerEndpoint;
+
+public class PhoenixRegionServerEndpointHAImpl extends PhoenixRegionServerEndpoint {
+    @Override
+    public ServerMetadataCache getServerMetadataCache() {
+        return ServerMetadataCacheHAImpl.getInstance(conf);
+    }
+}
