+++ phoenix-core/src/it/java/org/apache/phoenix/jdbc/HighAvailabilityTestingUtility.java
@@ -18,7 +18,10 @@
 package org.apache.phoenix.jdbc;
 
 import org.apache.hadoop.hbase.StartMiniClusterOption;
+import org.apache.phoenix.cache.ServerMetadataCache;
+import org.apache.phoenix.cache.ServerMetadataCacheImpl;
 import org.apache.phoenix.coprocessor.PhoenixRegionServerEndpoint;
+import org.apache.phoenix.end2end.PhoenixRegionServerEndpointHAImpl;
 import org.apache.phoenix.query.QueryServicesOptions;
 import org.apache.phoenix.thirdparty.com.google.common.annotations.VisibleForTesting;
 import org.apache.phoenix.thirdparty.com.google.common.collect.ImmutableList;
@@ -46,7 +49,9 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.Properties;
 import java.util.concurrent.Executors;
@@ -539,9 +544,9 @@ public class HighAvailabilityTestingUtility {
             conf.setInt(DFS_REPLICATION_KEY, 1); // we only need one replica for testing
 
             // Phoenix Region Server Endpoint needed for metadata caching
-            if (QueryServicesOptions.DEFAULT_LAST_DDL_TIMESTAMP_VALIDATION_ENABLED) {
+            if (QueryServicesOptions.DEFAULT_PHOENIX_METADATA_INVALIDATE_CACHE_ENABLED) {
                 conf.set(REGIONSERVER_COPROCESSOR_CONF_KEY,
-                            PhoenixRegionServerEndpoint.class.getName());
+                        PhoenixRegionServerEndpointHAImpl.class.getName());
             }
         }
     }
